boot/efibootguard: new boot package
This patch adds a new boot package to Buildroot to build the
EFI Boot Guard bootloader from the https://github.com/siemens/efibootguard
repository.

EFI Boot Guard can be chainloaded as an EFI application or used as a
standalone bootloader and can be useful when booting A/B systems.

Add a recipe to build the EFI bootloader application, user-space
tooling & host tooling (both of which are used to manage EFI Boot Guard
configuration).

EFI Boot Guard can be built for multiple architectures, but this initial
patch only builds the EFI bootloader for x86_64.

Signed-off-by: Christopher Obbard <obbardc@gmail.com>