board/pc-efibootguard: add example EFI Boot Guard recipe
Add an example recipe (based on configs/pc_x86_64_efi_defconfig) to
generate an x86-64 image which boots the system using the EFI Boot Guard
bootloader. The recipe is an example of how to create an image using
EFI Boot Guard, including how to create the disk image & how to configure
the bootloader.

The created example image contains two boot slots; one of which contains
the environment built by Buildroot and the other is an empty, dummy slot
(e.g. cannot be booted from).

EFI Boot Guard provides a tool for generating UKI images, which can then
be signed for Secure Boot. For simplicity, this is not enabled in the
example recipe.

Also add instructions (and a script) to boot the example image in QEMU.

Signed-off-by: Christopher Obbard <obbardc@gmail.com>